+++ doctorsAndPharmacy/src/main/java/com/rangers/medicineservice/configuration/ZoomRestTemplateConfig.java
@@ -1,15 +1,14 @@
-package com.rangers.medicineservice.config;
+package com.rangers.medicineservice.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
 
 @Configuration
-public class RestTemplateConfig {
+public class ZoomRestTemplateConfig {
 
-    @Bean
+    @Bean(name = "zoomRestTemplate")
     public RestTemplate zoomRestTemplate() {
         return new RestTemplate();
     }
-
 }
