+++ openai-chatbot/src/main/java/com/rangers/medicineservice/config/RestTemplateConfig.java
@@ -1,0 +1,25 @@
+package com.rangers.medicineservice.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class RestTemplateConfig {
+
+    @Value("${openai.key}")
+    private String openaiApiKey;
+
+    @Bean(name = "openai_restTemplate")
+    public RestTemplate restTemplate() {
+        RestTemplate restTemplate = new RestTemplate();
+        restTemplate.getInterceptors().add(
+                ((request, body, execution) -> {
+                    request.getHeaders()
+                            .add("Authorization", "Bearer " + openaiApiKey);
+                    return execution.execute(request, body);
+                }));
+        return restTemplate;
+    }
+}
