+++ openai-chatbot/src/main/java/com/rangers/medicineservice/config/TrainingDataConfig.java
